pos-backend/src/main/java/ar/com/terminal/dto/Item.java
package ar.com.terminal.dto;

/**
 * Created by ivan on 05/09/15.
 */
public class Item {

    private String id;
    private String description;
    private String price;

    public String getId() {
        return id;
    }

    public void setId(String id) {
        this.id = id;
    }

    public String getDescription() {
        return description;
    }

    public void setDescription(String description) {
        this.description = description;
    }

    public String getPrice() {
        return price;
    }

    public void setPrice(String price) {
        this.price = price;
    }

    public Item(){

    }
}
